StableCoin/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
bcrypt==4.3.0
bitarray==3.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
cytoolz==1.0.1
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.3.0
eth_abi==5.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
hexbytes==1.3.0
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
narwhals==1.27.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
paramiko==3.5.1
parsimonious==0.10.0
pillow==11.1.0
propcache==0.3.0
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyunormalize==16.0.0
pywin32==308
qrcode==8.1
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rlp==4.1.0
rpds-py==0.23.1
schedule==1.2.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sshtunnel==0.4.0
starlette==0.45.3
streamlit==1.42.2
tenacity==9.0.0
terminal==0.4.0
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
web3==7.8.0
websockets==13.1
yarl==1.18.3
